requirements.txt
python-dotenv
yara-python
requests